+++ packages/guide/src/app/locations/locations-shop-route-config.ts
@@ -13,6 +13,8 @@ import { ConcernedMonkeyComponent } from "./components/shops/concerned-monkey/co
 import { BosComponent } from "./components/shops/bos/bos.component";
 import { BensCaravanComponent } from "./components/shops/bens-caravan/bens-caravan.component";
 import { SocketAndPanComponent } from "./components/shops/socket-and-pan/socket-and-pan.component";
+import { WhiteFlamingoComponent } from "./components/shops/white-flamingo/white-flamingo.component";
+import { CoffeeComponent } from "./components/shops/coffee/coffee.component";
 
 export const shopRouteConfig: { name: ShopName, component: Type<any>, betaOnly?: true }[] = [
     {
@@ -65,5 +67,13 @@ export const shopRouteConfig: { name: ShopName, component: Type<any>, betaOnly?:
     {
         name: 'merfolk-oracle-tail-store',
         component: MerfolkOracleTailStoreComponent,
+    },
+    {
+        name: 'white-flamingo',
+        component: WhiteFlamingoComponent,
+    },
+    {
+        name: 'coffee',
+        component: CoffeeComponent,
     },
 ]