+++ packages/guide/src/app/locations/locations.module.ts
@@ -39,6 +39,8 @@ import { GenericShopComponent } from './components/shops/generic-shop/generic-sh
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { FestivalShopItemDetailsComponent } from './components/festival-shop-item-details/festival-shop-item-details.component';
 import { FestivalShopItemTableComponent } from './components/tables/festival-shop-item-table/festival-shop-item-table.component';
+import { CoffeeComponent } from "./components/shops/coffee/coffee.component";
+import { WhiteFlamingoComponent } from "./components/shops/white-flamingo/white-flamingo.component";
 
 @NgModule({
     declarations: [
@@ -73,6 +75,8 @@ import { FestivalShopItemTableComponent } from './components/tables/festival-sho
         GenericShopComponent,
         FestivalShopItemDetailsComponent,
         FestivalShopItemTableComponent,
+        CoffeeComponent,
+        WhiteFlamingoComponent
     ],
     imports: [
         CommonModule,